% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_distribution.R
\name{coef.fitted_distribution}
\alias{coef.fitted_distribution}
\title{Coefficients of fitted models}
\usage{
\method{coef}{fitted_distribution}(fit, studies, ...)
}
\arguments{
\item{fit}{A [PCNMA::fitted_distribution] object}

\item{...}{for S3 consistency}
}
\description{
Returns the model coefficients for a given survival model
}
