% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_distribution.R
\name{.fit_distribution}
\alias{.fit_distribution}
\title{Fit a single distribution}
\usage{
.fit_distribution(distribution, data)
}
\arguments{
\item{distribution}{A single distribution}

\item{data}{An IPD dataset}
}
\value{
A [flexsurv::flexsurvreg] object
}
\description{
Fit a single distribution
}
