% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_network.R
\name{gen_network}
\alias{gen_network}
\title{Generate a network of evidence}
\usage{
gen_network(net_data, ref, y = "RMST")
}
\arguments{
\item{net_data}{A dataset created by [PCNMA::gen_network_data]}

\item{ref}{A character reference treatment}
}
\value{
A [mutlinma::nma_data] object
}
\description{
Generate a network of evidence
}
