% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_network_data.R
\name{gen_network_data}
\alias{gen_network_data}
\title{Generate network data}
\usage{
gen_network_data(data, ref)
}
\arguments{
\item{data}{A date extraction dataset}

\item{ref}{A character reference treatment}
}
\value{
A dataframe
}
\description{
Generate network data
}
