% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fracpols.R
\name{H}
\alias{H}
\title{H-function for FPs}
\usage{
H(x, P, zeta, j)
}
\arguments{
\item{x}{A real value}

\item{P}{A vector of powers}

\item{zeta}{A vector of zeta values}

\item{j}{The index}
}
\description{
H-function for FPs
}
