% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/km_estimates.R
\name{hr}
\alias{hr}
\title{Hazard Ratio}
\usage{
hr(TTE, strata = "1")
}
\arguments{
\item{TTE}{A TTE dataframe}

\item{strata}{Strata for the RHS of the [survival::Surv] function}
}
\description{
Hazard Ratio
}
