% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/km_estimates.R
\name{km_estimates}
\alias{km_estimates}
\title{Generate KM estimates}
\usage{
km_estimates(TTE, strata = 1)
}
\arguments{
\item{TTE}{A TTE dataframe}
}
\value{
A [ggsurvfit::survfit2] object
}
\description{
Generate KM estimates
}
