% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fracpols.R
\name{phi}
\alias{phi}
\title{Fractional Polynomial Function}
\usage{
phi(x, m, P, zeta)
}
\arguments{
\item{x}{A real valye}

\item{m}{The degree of the polynomial}

\item{P}{A vector of powers}

\item{zeta}{A vector of zeta values}
}
\description{
Fractional Polynomial Function
}
