% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_distribution.R
\name{plot.fitted_distribution}
\alias{plot.fitted_distribution}
\title{Plot a fitted distribution}
\usage{
\method{plot}{fitted_distribution}(fit, km, ...)
}
\arguments{
\item{fit}{A [PCNMA::fitted_distribtuion] object}

\item{km}{a dataframe with a surv column}

\item{...}{For S3 consistency}
}
\description{
Plot a fitted distribution
}
