% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_distribution.R
\name{plot_fitted_distribution}
\alias{plot_fitted_distribution}
\title{Plot a fitted distribution}
\usage{
plot_fitted_distribution(fit, km)
}
\arguments{
\item{fit}{A [flexsurv::flexsurvreg] object}

\item{km}{a dataframe with a surv column}
}
\description{
Plot a fitted distribution
}
