% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_km.R
\name{plot_km}
\alias{plot_km}
\title{Plot a KM curve}
\usage{
plot_km(fit, break.x.by = 6)
}
\arguments{
\item{fit}{A [survival::survfit] object}

\item{break.x.by}{A numberic value for splitting x axis}
}
\value{
A plotted km curve
}
\description{
Plot a KM curve
}
