% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_distribution.R
\name{rmst}
\alias{rmst}
\title{Hazard ratios of a fitted model}
\usage{
rmst(fit, x, ...)
}
\arguments{
\item{fit}{A [PCNMA::fitted_distribution] objeect,}

\item{x}{Time to calculate RMST for}

\item{...}{For S3 consistency}
}
\description{
Hazard ratios of a fitted model
}
