% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_coefficients.R
\name{save_coefficients}
\alias{save_coefficients}
\title{Save coefficients of models}
\usage{
save_coefficients(models, study, treatment)
}
\arguments{
\item{models}{A [fitted_distribution] object}

\item{study}{The study}

\item{treatment}{The treatment}
}
\description{
Save coefficients of models
}
