% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_nma_summary.R
\name{save_nma_summary}
\alias{save_nma_summary}
\title{Write model summary to xlsx}
\usage{
save_nma_summary(nma, PARAMCD, effects, model_type)
}
\arguments{
\item{nma}{A [mutinma::stan_data] object}

\item{PARAMCD}{one of OS or PFS}

\item{model_type}{one of parametric or fp}

\item{effect}{type}
}
\description{
Write model summary to xlsx
}
