% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{summary.fitted_model}
\alias{summary.fitted_model}
\title{Summary of a  NMA model}
\usage{
\method{summary}{fitted_model}(model, dic = FALSE)
}
\arguments{
\item{model}{A `PancSurv::fitted_model` object}
}
\description{
Summary of a  NMA model
}
