% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/km_estimates.R
\name{summary.km_obj}
\alias{summary.km_obj}
\title{Summarise KM data}
\usage{
\method{summary}{km_obj}(fit, ...)
}
\arguments{
\item{fit}{A `PCNMA::km_obj` object}
}
\value{
A summary table of the KM data
}
\description{
Summarise KM data
}
