% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_coefficients.R
\name{write_coefficients}
\alias{write_coefficients}
\title{Write coefficients of models to excel}
\usage{
write_coefficients(models, study, treatment, PARAMCD)
}
\arguments{
\item{models}{A [fitted_distribution] object}

\item{study}{The study}

\item{treatment}{The treatment}
}
\description{
Write coefficients of models to excel
}
