% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_distribution.R
\name{fit_distribution}
\alias{fit_distribution}
\title{Fit survival distributions to a dataset.}
\usage{
fit_distribution(distributions, data, strata = "Treatment")
}
\arguments{
\item{distributions}{A list of distributions}

\item{data}{An IPD dataset}
}
\value{
A dataframe with fitted values
}
\description{
This function extends the `PCNMA:::.fit_distribution` function, by fitting a given set of distributions to a TTE dataset.
}
