% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_distribution.R
\name{fit_distribution}
\alias{fit_distribution}
\title{Fit distributions to a dataset}
\usage{
fit_distribution(distributions, data)
}
\arguments{
\item{distributions}{A list of distributions}

\item{data}{An IPD dataset}
}
\value{
A dataframe with fitted values
}
\description{
Fit distributions to a dataset
}
