% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{fit_model}
\alias{fit_model}
\title{Run an NMA}
\usage{
fit_model(network, L, link = "log", iter, seed)
}
\arguments{
\item{network}{A [multinma::nma_data] object}

\item{L}{Character string specifying a likelihood function}

\item{link}{Character string specifying a link function (defaults to "log")}

\item{...}{Other parameters to pass to [multinma::nma]}
}
\description{
Run an NMA
}
