% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/km_estimates.R
\name{km_estimates}
\alias{km_estimates}
\title{Generate KM estimates}
\usage{
km_estimates(TTE, strata = "1")
}
\arguments{
\item{TTE}{A TTE dataframe}

\item{Strata}{for the RHS of the [survival::Surv] function}
}
\value{
A [PCNMA::km_obj] object
}
\description{
Generate Kaplan-Meier estimates from a TTE object. Uses [ggsurvfit::survfit2] to generate KM estimates. Only requires strata to be specified.
}
