% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_distribution.R
\name{plot.fitted_distribution}
\alias{plot.fitted_distribution}
\title{Plot a fitted distributions object}
\usage{
\method{plot}{fitted_distribution}(
  fit,
  CI = FALSE,
  km = FALSE,
  km_alpha = 1,
  linewidth = 0.75,
  linetype = "dashed",
  theme = "bw",
  facet_by = "Treatment",
  ...
)
}
\arguments{
\item{fit}{A `PCNMA::fitted_distribution` object}

\item{CI}{Include a confidence interval?}

\item{km}{Add the original KM curve?}

\item{...}{For S3 consistency}
}
\description{
Creates a plot for a result of `PCNMA::fit_distribution`.
}
