% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/km_estimates.R
\name{plot.km_obj}
\alias{plot.km_obj}
\title{Plot a KM curve}
\usage{
\method{plot}{km_obj}(fit, break.x.by = 5, xMax = 40, risktable.height = 0.3, ...)
}
\arguments{
\item{fit}{A [PCNMA::km_obj] object}

\item{break.x.by}{A numeric value for splitting x axis}

\item{xMax}{The maximum time value to plot}

\item{risktable.height}{The proportion of the figure to be taken up by the risk table}

\item{...}{For S3 consistency}
}
\value{
A plotted km curve
}
\description{
Plot a KM curve
}
