% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_distribution.R
\name{rmst}
\alias{rmst}
\title{RMST of fitted survival models}
\usage{
rmst(fit, x, ...)
}
\arguments{
\item{fit}{A [PCNMA::fitted_distribution] object}

\item{x}{Time to calculate RMST for}

\item{...}{For S3 consistency}
}
\description{
Returns the RMST estimate for a fitted survival model up to a time x.
}
