% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_distribution.R
\name{summary.fitted_distribution}
\alias{summary.fitted_distribution}
\title{Summary of a set of fitted models}
\usage{
\method{summary}{fitted_distribution}(fit, AIC = FALSE)
}
\arguments{
\item{fit}{A `PCNMA::fitted_distibutions` object.}
}
\description{
Summary of a set of fitted models
}
