% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{summary.fitted_model}
\alias{summary.fitted_model}
\title{Summary of an NMA model}
\usage{
\method{summary}{fitted_model}(model, likelihood, effect)
}
\arguments{
\item{model}{A [PCNMA::fitted_model] object}

\item{likelihood}{A vector of likelihoods}

\item{effect}{FE or RE}
}
\description{
Summary of an NMA model
}
